chatbot_project/requirements.txt
faiss-cpu
transformers
torch
requests
numpy
groq
